requirements.txt
django
django-tinymce4-lite
django-flat-responsive
mysqlclient
